refactor: combined registration and login in 1 controller
